mcp: carve controller_loop.rs (pure move)
The controller-restart machinery and controller-loop/wrapper
observability leave mod.rs: restart state machine + validation +
persistence, loop-status collection from run dirs, external-wrapper
index and live codex-app-server process probing, halt/intervention
markers, and the scheduled-restart runner — 1,996 body lines plus their
34 tests (1,482 lines) verbatim.

Move glue per the split recipe: `use super::*;` inherits mod.rs's
import view; moved items pub(crate)-widened (plus
ControllerLoopIntervention's two fields, accessed from mod.rs);
`super::X` paths re-anchored to the absolute forms they already
resolved to (`crate::X` in bodies, `crate::mcp::X` in tests — the
module is one level deeper now); the shared test fixture
test_state_with_log_dir widened for cross-module use. Verified by the
toolkit purity check; mcp:: tests 178/178 unchanged.
